Successful refactoring to follow MVC.
